Replace Double with Float in DSPUtils, MeterState, SpectrumEngine, AnalogVUMeterView, LevelMeterView, MiniPlayerContent, and SpectrumView for consistency

+++ Sources/CamillaDSPMonitor/Models/DSPUtils.swift
@@ -5,8 +5,8 @@ import Foundation
 // MARK: - Shared Audio Utilities
 
 /// Normalize a dB value (-60..0) to 0..1 range for meter/spectrum display.
-func normalizedDB(_ db: Double) -> Double {
-  max(0, min(1, (db + 60) / 60))
+func normalizedDB(_ db: Float) -> Double {
+  max(0, min(1, (Double(db) + 60) / 60))
 }
 
 private let _rateFormatter: NumberFormatter = {
